update: organize namespacing and imports
