+++ src/components/AboutHeader/IAboutHeader.ts
@@ -1,5 +1,5 @@
 import {ComponentProps} from 'react'
-import {SocialLink} from '../../types/SocialLink'
+import {SocialLink} from '../../types'
 
 export interface IAboutHeader extends ComponentProps<'header'> {
     links: SocialLink[]
