+++ src/components/Hero/index.ts
