+++ src/components/ProjectCard/IProjectCard.ts
@@ -1,4 +1,4 @@
-import {Project} from '../../models'
+import {Project} from '../../entities'
 import {ComponentProps} from 'react'
 
 export interface IProjectCard extends ComponentProps<'article'> {
