+++ src/components/ProjectsList/IProjectsList.ts
@@ -1,5 +1,5 @@
 import {ComponentProps} from 'react'
-import {Project} from '../../models'
+import {Project} from '../../entities'
 
 export interface IProjectsList extends ComponentProps<'ul'> {
     projects: Project[]
