+++ src/components/SkillsList/ISkillsList.ts
@@ -1,5 +1,5 @@
 import {ComponentProps} from 'react'
-import {Skill} from '../../models'
+import {Skill} from '../../entities'
 
 export interface ISkillsList extends ComponentProps<'ul'> {
     skills: Skill[]
