+++ src/components/SkillsList/index.ts
@@ -1,1 +1,2 @@
 export * from './ISkillsList'
+export * from './SkillsList'
