+++ src/components/index.ts
@@ -10,6 +10,8 @@ export * from './ProjectsList'
 export * from './ProjectsHeader'
 export * from './SkillsPanel'
 export * from './SkillCard'
+export * from './SkillsList'
+export * from './SkillsHeader'
 export * from './ProjectCard'
 export * from './ProjectFilter'
 export * from './Pitch'
