+++ src/context/MainContext/index.ts
@@ -1,1 +1,2 @@
 export * from './MainContext'
+export * from './IMainContext'
