+++ src/context/ProjectsContext/index.ts
@@ -1,1 +1,2 @@
 export * from './ProjectsContext'
+export * from './IProjectsContext'
