+++ src/context/SkillsContext/index.ts
@@ -1,1 +1,2 @@
 export * from './SkillsContext'
+export * from './ISkillsContext'
