+++ src/entities/Entity.ts
