+++ src/entities/Skill.ts
