+++ src/pages/About/IAbout.ts
@@ -1,5 +1,5 @@
 import {ComponentProps} from 'react'
-import {SocialLink} from '../../types/SocialLink'
+import {SocialLink} from '../../types'
 
 export interface IAbout extends ComponentProps<'article'> {
     links: SocialLink[]
