+++ src/services/project-service.ts
@@ -1,4 +1,4 @@
-import {Project} from '../models'
+import {Project} from '../entities'
 
 function getPlaceholderProjects(): Project[] {
     return [
